Rewrote piece shuffling algorithm to be more reliable.
The old algorithm shuffled pieces up to 48 times and hoped that one of the
random arrangements resulted in having no back-to-back pieces. But it was
random, and even though the most I ever saw it take was 15-20 tries, it
could theoretically have bad luck and fail to come up with an arrangement.

The new algorithm finds a location for each piece deterministically, and
is more reliable. It will only fail if it's impossible to find an
arrangement (e.g a custom level with a piece queue with 3 T pieces and an
O piece)