Slower tutorial dialog. Unselected chat choices are more faded.
The player has about 1.5 extra seconds to read tutorial text, and the demos
are slowed down as well.

When playtesting the demo, testers reported that the text was hard to
read, and that it was difficult to read and watch a demo at the same time.
They requested that we only do one at a time, and that we have an
'advance' button to advance the text. While I agree that would allow for
the greatest versatility between very fast/very slow readers, it also
seems like it requires more button presses/interaction, and I'd rather try
something like this first.

Decreased transparency of unselected chat choice when presenting the
player with a set of choices. During playtests, testers had trouble telling
which of the two choices was selected.

Closes #703. Closes #704.