Fixed bug where all chat icons blinked sometimes.
When the player runs near different creatures, their chat icon is supposed
to light up. But when this happened, all of the chat bubbles blinked. This
is because the tween's parameters were using the wrong node's modulate
proeprty.

I also added a check to avoid launching unnecessary chat icon tweens. This
shouldn't have any visible effect, but fewer tweens is good for performance and
simplicity.