Extracted LevelLibrary. Fixed possible 'inaccessible tutorial' bug.
Other classes and logic need to know things like 'which levels does the
player have unlocked' outside of the level select UI, and it makes sense
for this to be a singleton node so we don't need to reload the JSON
everywhere.

Fixed potential tutorial bug. There was some logic setting the tutorial world to
LevelLock.STATUS_CLEARED (1) if all levels in the tutorial were complete.
However, tutorial worlds don't use this particular enum, and it resulted in the
tutorials being marked as LockStatus.LOCK (1) instead.