Fixed bugs in overworld dialog, limbs
Fixed bug with the wrong dialog showing on the overworld when you selected
the 'level 1'/'level 2'/'chit chat' options.

Fixed bug with facial expressions no longer working.

These two bugs were caused because ChattableManager was loading all of the
creatures on _ready(), but it's an autoload node which will only be loaded
once. This meant it would work OK if Overworld.tscn was loaded directly,
but didn't work when the game was launched normally through the title
screen.

Closes #750.

Fixed bug where arms/legs pointed the wrong way if you talked to a character
walking backwards. This is because the 'idle-nw' animation overrode the newly
set arm and leg frames.

Closes #751.