Added restaurant to overworld. Fixed north-facing bugs.
There are arrows for entering/exiting the over world. Stepping on the
arrows transitions the player to a new area. When the player exits the
restaurant, they spawn at different overworld spawn points.

I experimented with having the restaurant be an isometric resource but
this causes rendering order problems. Two obstacles with the same
Y-coordinate might need to be rendered in front of and behind the
restaurant, based on their X-coordinate. It's better to have all obstacles
face forward, or to have the same width/height.

Enabled 'edit group' for many overworld obstacles.

Fixed bug where creatures with an initial drection of northeast would have
their bodies/heads flipped incorrectly.

Extracted OverworldObstacles, OverworldDemo. This allows us to see the
entirety of the overworld tilemap as we're editing it, as opposed to only
seeing the tiles which are in the special goop viewport.