Overworld refactorings; OverworldUi scene, simpler paths.
Nodes now use OverworldUi dependency instead of OverworldUiLayer dependency.
This dependency is referenced with a global utility function
(Global.get_overworld_ui) instead of complex relative paths (../../../Ui)

ChattableManager.clear() is now handled with a breadcrumb listener. Ideally this
would be with a scene_changed signal on the node tree, but the node tree does
not provide signals like that.

Fixed oversight where sensei was not purged from ChattableManager.