Initial cutscene location logic. Level dialog is no longer on creatures.
Cutscenes can take place indoors or outdoors. Added logic for spawning
cutscene characters at different locations.

Level dialog used to be attached to creatures who had specific logic for
'use this dialog for level 1, this dialog for level 2.' This logic is now
implicit, and level dialog is in /assets/main/puzzle/levels/cutscenes

Levels no longer have 'level numbers'. A creature used to have the idea of
their first and second level. This is now implicit based on the contents
of worlds.json and how the levels are organized.

Creature clipping is disabled during cutscene. Very, very, very fat
creatures will inevitably break cutscenes a little but it's better for it
to be cosmetic stuff (creatures clipping or looking funny) and not
functional stuff (creatures spawning out of bounds or not being able to
run somewhere)