Fix bug where dialog bubbles didn't appear on the correct side.
The logic for dialog bubbles appearing on one side or the other was broken
in 9c4c4b2f when the get_orientation function was removed. The code didn't
throw errors because ChatUi tried to invoked this function leniently using
reflection, instead of enforcing the chatter's class.

ChatUi now assumes all chatters are creatures.