Added creature stress test. FPS displayed as integer.
Added creature stress test which allows developers to measure the
performance impact of adding more and more creatures. This stress test
can be launched from the main menu in debug builds.

get_frames_per_second returns a float, but it is always an integer
value. It was unnecessary to always show '60.00 fps' so now this is
instead shown as '60'