Added support for localization using gettext.
A proof-of-concept spanish translation is in es.po; it is an incomplete
machine translation. But it demonstrates we can translate tutorial
dialog, character dialog, cutscenes, level information, messages,
buttons, and labels.

There is a new tool 'LocalizationDemo' which extracts some of these
strings for use in the PO template. More information on this tool is available
in the turbo fat wiki.

Embedded SplashScreen's button in a CenterContainer to ensure it stays
in the center when its text is changed as a result of localization.