Remove lull characters from dialog, tutorials.
These lull characters are tedious to type, and interfere with
translation. A translator will understand how to translate 'Well?
...What is it?' to japanese. A translator will not understand how to
translate 'Well?/ ./././What is it?' to japanese.

The only line which kept its lull characters is the 'pap-pap' line
because rhythm is important to it. There may be other lines like this in
the future.