Removed overworld goop/ground viewports.
The overworld used to have two viewports for rendering the goop and
ground. It would render the goop to one ViewportTexture, then use
combine that texture with a few traditional image-based textures to
render the ground.

Instead, it now applies all of this rendering logic in a shader on the
tileset. It does not require any viewports.

Surprisingly, this does not have a significant impact on performance.
There is a performance benefit in avoiding the two viewports, but a
performance cost in applying a shader to 500+ tiles.

Closes #826. Closes #742.

Increased the cell_quadrant_size for the overworld. The goop shader has
visible seams at the quadrant edges, and I couldn't find a way to avoid
this. The easiest fix is simply to avoid having goop tiles overlay
seams.

Removed OverworldObstacles, OverworldDemo. These were only kept separate
because of the difficulty of working with the viewport-based rendering
approach.

Resolved a possible NPE when rendering a body before its shadow node
exists.