Fixed localization bugs: tutorial, fuzzy match
Fixed 'mega_pause_charaters' typo. This method was renamed to
mega_lull_characters but the reference was not changed. The result was
that triggering the tutorial easter egg would make the tutorial
unfinishable.

Fixed bug where unsupported locale caused menu to be inconsistent.
Specifically, when testing the game my locale was 'en_US' but this did
not match our supported locale of 'en', so it set the selected language
menu item to 'Spanish' even though the game was displayed in english.

The game now matches 'en' to the locale 'en_US', and 'es' to the local
'es_MX' in the settings menu. It does this with an algorithm similar to
the algorithm used by Godot for locale matching.