Fix ghost piece toggle bug, splash fade bug.
Toggling the ghost piece in the middle of a puzzle caused some unusual
behavior. The shadow offset would retain the value for the most recent
shadow piece. So the ghost piece stuck around, it was just in a silly
place. We now enable/disable the ghost piece correctly and respond to
the player changing their settings mid-puzzle.

Closes #853.

Fixed a bug where clicking 'Play' while the splash screen was still
fading in resulted in all the subsequent menus fading in and out as
well.