Renamed survival to marathon.
I renamed Marathon to Survival in 47e2c61c (2020-05-29) with the justification
that it was the only mode where topping out ends your game, so 'survival' was a
more fitting name. However, the Turbo Fat community has pointed out that
'Survival' is misleading as you're evaluated on score, not on how long you
survive.