Replaced json dialog with chatscript.
Converted all dialog and cutscenes from json to chatscript. Chatscript
is a condensed format which is less versatile but much easier to write.

Closes #782.

Removed support for creatures with inline dialog.

Pulled bort's dialog out of his creature.json. I think this may have
been a proof of concept for the 'all in one file' type creature, which
may eventually be supported by the creature editor. But right now it's a
lot of complexity for complexity's sake -- if we go the 'all in one
file' approach we'll need to think about this case then.

load_chat_events -> load_chat_tree; 'ChatEvents' was renamed to
'ChatTree' a long time ago but this one reference was not changed.

Removed unused 'no-vegetables', 'level-1', 'level-2' conversations.
These were redundant with the cutscenes and never appeared in-game.