Enabled vsync. Food NPE fix. Removed wallpaper kludge.
Added fix for NPE when food items were deleted at unexpected times

As of Godot 3.3, this wallpaper sprite bug does not appear to be an
issue. Or, it's transient and I can't figure it out. It happened this
morning after the 3.3 upgrade, but after restarting Godot it hasn't
happened again -- even if I remove the invisible textlabel which was the
workaround for older Godot versions.