Bunny ear fixes, enabled_focus_mode, suppress_fatness.
Fixed new bunny ear sprites which were missing buffer area, causing
visual smudges.

Removed deprecated 'enabled_focus_mode' property. This was producing
warnings because it is redundant with another property and will be removed in
Godot 4.0.

Toggled suppress_fatness for sensei/player on overworld. Before, feeding a
creature with the same name as the player was causing the player to gain weight.