Removed support for May 2020 save files.
Save files from May 2020 (0.0517) were very different. They used a
different filename, were not in JSON, and did not have a version number.
Removing support for this format lets us get rid of a lot of special
cases for handling this separate filepath and translating between JSON
and non-JSON formats.