Improve starting customer logic, food targeting efficiency
The old starting customer logic had conflicting logic in the RestaurantView and
Puzzle scripts. 'Puzzle' would reload all the new creatures, and then
'RestaurantView' would try to scroll to a new creature if they had been fed, but
since all creatures had just been reloaded, it would always just scroll to a new
creature.

The new starting customer logic is more robust, and handles edge cases
such as restarting a puzzle when no creatures have been fed, restarting
a story puzzle when the camera is on the starting creature, and so on.

FoodItems target calculations are now cached, instead of being
recalculated 20 times for 20 different food items. They're all going to
the same location so it's silly to perform multiple lerps/transforms
over and over to get the same answer.