Added houses to the environment. Skip caching overworld resources.
The overworld now includes 13 different residences. These are each
contained in their own scene. I considered making a generic scene which
would let you set something like the texture and offset of each house.
However, almost every object in the overworld is just going to be a 2D scene
with a collision shape, and a sprite. It didn't seem reusable.

ResourceCache does not load environment PNGs up front. Images for the different
houses and locations on the overworld are large. These increase the memory
footprint from ~170 MiB to ~210 MiB and as the game grows, only some of them
will be visible at any given time. It doesn't make sense to cache all of these
up front.