Changed indoor floor tiles from rectangles to parallelograms.
Having the interior tiles and collision shapes as rectangles resulted in a
problem when trying to design a vertical countertop. This would either need to
be diagonal and not align with the tilemap, or it would need to be vertical and
drawn strangely.

This affects the indoor floor tiles, shadows, tilemap and collision
shapes. Assets such as tables, chairs and walls were not redrawn. Table
and chair collision shapes remain as rectangles, as this feels more
intuitive for now.