Cutscene camera, orientation fixes.
Moved cutscene camera lower. The dialog bubbles were hiding the creatures.

Fix cutscene camera bug for levels with no creature. The 'no creature'
creature would be spawned at a default location and the camera would
struggle to keep them in frame.

Creatures no longer turn left and right during cutscenes. This was
resulting in unnatural behavior where, in a group of 5 creatures, they would all
face the player even when someone else was talking.

The sensei and player orient correctly during conversations between the player
and the sensei. The 'calculate the midpoint of the chatters' logic was not
behaving correctly when they were the only ones talking.

Sensei no longer follows player during cutscenes

Fixed bug where cutscenes would dump player back to the title screen
after a puzzle, if the cutscene was marked as 'skipped'.