Merrymellow Marsh sequence changes.
Added prioritized_if/skip_if flags to Merrymellow Marsh levels. The
cutscenes are prioritized when they're next in sequence, and skipped if
the player's past that point in the story.

Level sequence changes; only need to complete about 70% of the
Merrymellow Marsh levels

Set creature id to richie for marsh cutscenes. Without a creature_id,
these cutscenes cannot be accessed outside of the phone menu.