Indoor isometric tiles are sorted with a transform.
By applying an unnoticable y component to the tilemap transformation,
tiles on the right are rendered in front of tiles on the left. 0.007
appears to be the smallest value which works. Values smaller than that
result in the tiles being left unsorted.

This is a workaround for the absence of x-sorting in tilemaps, a feature
requested by Godot proposal #2838
https://github.com/godotengine/godot-proposals/issues/2838