Smoother cutscene cameras. Better camera adjustments for fat creatures.
Creatures entering or exiting a cutscene now launch a tween which smoothly moves
the camera instead of lerping. The previous lerp effect caused the camera to
leap into motion which was unpleasant.

Cutscene camera zooms and pans slower than before. The rapid zooming and
panning was also unpleasant.

Better camera adjustments for fat creatures. Camera is adjusted vertically to
keep fat creature's faces in frame. Camera is zoomed out when fat creatures are
in a cutscene, this logic was accidentally removed in f0f31d1

hello-everyone-000 cutscene no longer fades in three creatures on the
first line of dialog. This was unnoticable, but resulted in the camera
(correctly) zooming in on just the two visible creatures, and pulling out to all
five creatures, which looked strange.