Delint improvements: headers, grep performance
Improved delint script so that when executed, it precedes each issue
with a header like 'Functions missing return type:'

Improved performance of delint script. The 'long lines' detection was
taking awhile to run because it executed six recursive grep statements
consecutively. It now runs a single grep statement with a more
complicated expression.