Replaced preroll/postroll logic with CutsceneManager
CutsceneManager provides the idea of a queue of cutscenes. This allows for
cutscenes which aren't follow or preceded by a level.

CurrentLevel no tracks the 'cutscene state', but instead tracks whether
the player has lost/finished/won the current level. It uses this state
to decide whether or not to load the postroll cutscene into the
CutsceneManager following a puzzle.