Replaced ambiguous 'chat' prefix with 'creature', 'level'
Before, the 'chat' prefix could be used to point to resources in
/assets/main/creatures/primary or /assets/main/chat. These are now
referenced by different prefixes, 'creature' and 'chat', and there is a
third path 'level' for levels.

Replace hyphens with underscores in creature library. Creature library
used hyphens for '#fatness-023#' and a long time ago, for a creature
with an id of 'i-n-cognito'

CreatureLibrary.reset() resets fatnesses. This is necessary for unit
tests.