Removed PieceQueue autoload singleton.
PieceQueue was an autoload singleton to support LevelTriggerEffects,
which needed to rotate the pieces in the queue. However this sets a
difficult precedent for future effects which need to change the
active piece or add rows to the playfield. We don't want autoload
singletons for the active piece, playfield, and a hundred other
puzzle-specific things.

I've changed the design so that CurrentLevel keeps a reference to the
Puzzle scene, which contains a PieceQueue. When a puzzle is active,
LevelTriggerEffects can go through CurrentLevel to obtain references to
its piece queue, active piece, playfield, or a hundred other
puzzle-specific things.