Creature/level editor path refactoring. Removed unused code.
When importing/exporting files, the creature editor and level editor now
remember the path which was imported and recommend to export to the same
path. This way if you want to import from File A and export to File A,
you don't need to navigate to it twice.

The level editor's 'level_key' field is no longer editable. It is for
display purposes only, and shows the path the level was loaded from.

Removed unused level editor quit code. The level editor had some edge-case code
when quitting a level. This code never seemed to execute, and was causing an
error 'Can't take value from empty array.'