Player's creature_id is always '#player#'.
The player's creature_id in the save data corresponded to their name like
'gordo' or 'dusty' which could cause problems. It meant that other creatures in
the world could share their creature id. It meant the player could 'feed
themselves' and get fat.

In the spirit of refactoring week, it also meant that Creature.gd needed
some clumsy code to ask 'wait, are we loading the creature'. The
resulting code is now cleaner, as the special case code is pushed out
into special places.