Added pickups. Minor refactoring.
Pickups are seeds/stars on the playfield which can be collected. When collected,
they turn into foods which score the player points and keep their combo active.
Currently, no levels feature pickups but they can be shown in the PuzzleDemo
with the 'P' key.

Fixed a few comments. Added a comment explicitly stating why we remove items
when the game is paused, since it was unclear whether this was to avoid visual
clutter or visual bugs. Fixed Frosting Globs comment which incorrectly stated
that globs were pooled -- globs are no longer pooled.

Moved wobbler scale logic. The food item and wobbler scales were out of sync,
one had a base scale of 0.5 but the other had a base scale of 1.0. They're now
the same.

Food item jiggle is no longer an autoplay animation. It is now triggered
manually via the jiggle() or collect() methods.