Pickup improvements: renames, pickups at start
Renamed pickup methods to match other similar classes.

"set_pickup" is a better name than "add" because the pickups replace other
pickups in the same spot, and it is similar to "set_cell"

"clear" matches other similar classes

"remove" is better as a public method. Nobody uses it yet but its
intuitive behavior could be used in a demo or on a specific level.

Pickups are now spawned in the _ready() method so they will appear
when the puzzle scene first starts up.