Pickups go behind playfield blocks. FoodItem tweak.
This required increasing the Z index for many puzzle components by 3.
The z-index of pickups is '4' to go ahead of the active piece. But tiles
need to overlap pickups, and other UI items need to overlap tiles.

FoodItem no longer assigns its rotation/scale on 'ready'. This was
causing unnecessary changes to FoodItem.tscn and Pickup.tscn