Level editor calculates pickup score. Level editor drag/drop shortcut.
Level editor's 'properties' tab now has a panel for the pickup score.
The player can enter a score, or click the 'calculate' button to sum up
all of the visible pickups.

Dragging from the level editor playfield will now repeat the previously
dropped item for convenience. This makes it easier to create puzzles
with tons of veggies or powerups.