Doubled default pickup points. Rebalanced levels.
On a default level, an experienced player can earn ¥1000 per minute. But
on other levels, even with tons of powerups, it is difficult to score
more than ¥500 per minute. I want powerups to offset the disadvantage of
harder levels, but it is hard to do that unless they are worth more
points.

Fixed a calculation issue in LevelRankDemo. Rebalanced level rank data
for new powerup levels.