Fixed save corruption bug with pickups. Vee kick tests.
Fixed bug where rank calculator calculated a value of '1#INF' which got
converted to JSON, but couldn't be converted back from JSON. This
occurred when calculating the log of zero. The rank calculator now
clamps the resulting rank to a value of 999.

Added tests for vee kicks which I want to implement -- but which
conflict with wall kicks and snack kicks.