Created release toolkit. Added level upgrade logic. Localizable bug fix.
Created a centralized release toolkit. This includes the old 'LocalizationDemo'
but can expand to include any other utilities that are a part of the release
process. Having this centralized will prevent us from needing to memorize the
location of a dozen different demos.

Added level upgrade logic. This searches for levels with old version
data and upgrades them to the new version. This is important for two
reasons. First, we may eventually drop support for older level data.
Secondly, in the short term the level editor has no understanding of the
version tag, and a player might open a level, edit it and save it. This
would result in a level with a combination of old and new data formats.

Moved LEVEL_DATA_VERSION from level-settings.gd to levels.gd to avoid a
circular reference. Extracted level settings upgrade logic from
level-settings.gd into level-settings-upgrader.gd for reuse by the
release toolkit.

Fixed a bug where the 'Extract Localizables' button in the
LocalizationDemo would duplicate the scancode localizables if pressed
multiple times.