Refactoring: Upgrader method/class renames
Renamed classes like 'OldPlayerSave' and 'OldSystemSave' to
'PlayerSaveUpgrader'. This name makes it clear that this class is a verb
(it does something) as opposed to a noun (it has something.)

Replaced verbs like 'convert' and 'transform' into 'upgrade'. Upgrade is
more specific. These conversions are always one way; they only ever
upgrade to a newer version.