Fixed bug where inserted lines did not work.
This stems from a type conversion error which affects released versions of the
game. When performing float operations, Godot will haphazardly store the result
as a float or an int depending on the circumstances. In the Godot editor, it
likes storing them as ints more often. But when exported, it likes storing them
as floats more often.