Improved V piece climb kick.
A V piece will now climb up a one-tile stairwell instead of being pushed
off to the left.

A twitch streamer 'PrincessMomo13' attempted this kick and was surprised
the piece wouldn't climb in this way, and I agreed it was surprising.

Renamed misnamed 'climb_2l' test, which was actually rotating the piece
from '2' to 'r'.