Overworld offset, collision fixes. map_to_world_centered() utility.
Minor adjustments to overworld texture and collision offsets. The
empty outer cell collisions were slightly misaligned with the obstacle
collisions. The ground and obstacle tilemaps were also not very well
aligned with the tilemap grid. This became especially apparent when
adding goop waves because they were asymmetrical on the left and right.

Fixed some overworld tilesets which did not have collision shapes
defined. This was broken when we recently got rid of the goop shader and
generated pre-colored tilesets.

Extracted Utils.map_to_world_centered() utility. We were doing this in
many traditional tilemaps, but the procedure for calculating the center of a
cell in an isometric tilemap is more complex.