Utilize ResourceCache to load creatures faster.
ResourceCache was previously only used for caching large scenes to
reduce load times during scene transitions. It was already caching things like
images and sounds to facilitate in the caching of scenes, but it was
never polled for these resources directly.

The CreatureLoader and Creature classes now directly consult
ResourceCache to obtain creature resources. This is much faster and
eliminates the jitter effect when creatures show up during puzzles.