Refactorings: Test comments, repeat upgrades, piece signals.
Removed class comments from tests. Most of these test comments were completely
redundant with the comments on the target class. Even the more meaningful
comments didn't have much to say, so I've removed almost all of them.

SaveItemUpgrader upgrade logic now repeatedly upgrades data until it's
at the newest version, instead of only upgrading it one step. This is
identical to LevelSettingsUpgrader's logic.

Signal renames; playfield_disturbed, piece_disturbed

The old 'piece_changed' signal name was misleading because most of the
'score_changed' 'orientation_changed' values are emitted when the value is
replaced with a new value. piece_changed was being emitted for other cases, like
when the piece was moved or rotated. 'piece_disturbed' was also slightly
misleading because it was emitted in response to changes in the playfield, not
changes in the piece.