Fixed bugs related to creature weight during puzzles.
In levels ending after a fixed number of customers, the final customer
would get skinny again during the final sequence of line clears. This is
because their fatness was based on the bonus score, but the bonus score
decreased for those kinds of puzzles. Their fatness is now based on a
separate 'fatness score'.

During sequential line clears, the final foods eaten added a lot more
weight than the earlier foods. Imagine a series of ten simultaneous line clears,
each worth 35 points. The creature's weight for each line could be calculated as
adding 10% of the total score, or 100% of the line score. However the bug
occurred because the calculation was doing both; early line clears were lerping
10% of 35 points, but late line clers were lerping 90% of the way towards 350
points. I've fixed the calculation to work line-by-line, ignoring the
overall score.