Added 'preplaced pieces' rank logic
Some levels such as 'Teensy Taste' had boxes or pieces already on the
playfield. This meant the game would award an M rank for a time of 0:16 seconds
because it thought the player needed 30 pieces to finish the level. However it
should require a time of 0:12 seconds for an M rank, because the player only
needs 25 pieces.

The 'preplaced pieces' value for each level factors in how helpful the
pieces are. On Teensy Taste for example, there are five pieces on the
playfield, but preplaced_pieces is deliberately set to four, because the
boxes are arranged in a way that requires slow play.