Added career meter to puzzles. Tightened step calculation.
The puzzle now shows a meter next to the player which indicates how far
they'll travel. This meter increases (and sometimes decreases) as they
progress through a puzzle.

The step calculation now makes it harder to travel 2, 3 or 4 steps.
Before, it was incredibly trivial to travel 2 or 3 steps -- it could
often be done in 2-3 line clears.