Prevent save scumming of career mode boss levels.
Career mode advanced the clock and set distance earned to '0' to
discourage savescumming -- but for boss levels this isn't enough,
because they actually apply a penalty on failure. Career mode now
applies this penalty before the level starts.

Renamed 'advance_distance' to 'advance_clock' and changed its behavior
slightly. This allows us to reuse it for the savescumming logic.