Fixed career mode win transition bugs
The transition to Career Mode's win screen was broken when introducing
the fade in effect. This is because the CareerMap scene would apply a
fade in and fade out scene transition simultaneously, resulting in an
empty CareerMap screen loading.

Fixed bug where CareerWin did not remove singletons, resulting in errors
when the singleton MusicPopup was freed.

Fixed 'Class Instance Is Gone' errors in SceneTransitionCover if the
scene changes while fading in.