Refactoring: Raw json ints, Vector2.ZERO
All ints/floats from json are explicitly converted to ints/floats. We
have had several surface from not casting explicitly. These values are
stored as floats in the editor, but stored as ints when the game is
exported, resulting in confusing bugs. For example, last week the clock
times were displayed as '?:?? zm' because the released game was trying
to use 5.00 as an integer key.

Switched some code referencing 'Vector2(0, 0)' to the more succinct 'Vector2.ZERO'. I left it alone in places where it appeared alongside other numeric values.