CareerWin fixes: SceneTransitionCover, Control.
CareerWin had a redundant SceneTransitionCover. This caused redundant
calls to BreadCrumb and bugs in ResourceCache. However, these bugs only
seemed to occur when CareerWin was launched standalone, and don't appear to
affect the released game.

Changed CareerWin from a Node to a Control.

When CareerWin was a Node, a baffling side effect occurred. When
launched as a standalone scene, if you clicked 'OK' and navigated to the
main menu, the main menu's wallpaper would overlay its buttons. Changing it from
a Node to a Control fixes this somehow.