Refactor: Money icon, add_obstacle, rank milestones
Money icon is in a control with a nested sprite to avoid UI issues with
scaled controls.

Added ObstacleManager methods for adding obstacles.

Consolidated rank milestone logic from puzzle.gd and step-meter.gd into
career-data.gd. This includes color data which is a questionable choice
as it's UI specific, but this way is less error prone than having a
separate array for UI data.