Fix temporary status bar visual glitch when selecting a level.
After selecting a level in Career mode, the status bar would briefly
flash from '53 (+3)' to '52 (+1)' because the distance label would
detect the distance_travelled change signal. However, the
distance_travelled change was applying a penalty which the distance
label display already displays -- so this value was incorrect.

CareerMap now includes an explicit step to overwrite this label with the
correct value, so it flashes from '53 (+3)' to '53'