Career mode keeps level buttons focused
Previously in career mode, when clicking the settings button and then clicking
'OK', the level buttons would be deselected. It now reselects the previously
selected level button.

Previously in career mode, when clicking to choose a new start level,
the level buttons would be deselected. It now reselects the previously
selected level button.