Career cutscene ordering. Marsh career cutscenes.
Added career cutscene ordering logic. Numbered cutscenes like '00',
'01', '02' are played in order. Lettered cutscenes like 'a', 'b', 'c'
are ordered randomly. Cutscenes are arranged in a tree and these
ordering rules apply both to branches and children.

Added career cutscene selection logic. Each region can now define a
cutscene_path which points to a folder of cutscene files.

Added merrymellow marsh career cutscenes. Rewrote some dialog to account
for the fact that the player is not confined to Merrymellow Marsh in
Career mode, in the same way they would be in story mode. So instead of things
like 'We're leaving' the dialog now talks about 'We won't be visiting as much'.

ChatLibrary now properly handles resources in 'src/test' as well as
resources in 'src/main'.