SceneTransition handles 'don't fade out while fading in' scenario
Instead of CareerData and other places calling Breadcrumb directly to
avoid the 'don't fade out while fading in' scenario, they all delegate
to SceneTransition which handles it internally.

Consistency changes for SceneTransition. It now assigns 'fading = false'
for all of its transitions instead of just one, and handles the '::'
paths for all of its transitions instead of just one or two.

Minor fixes to comments