Add diota's career scenes. Career button can launch cutscene.
Added diota's scenes to career mode.

Added logic for career mode to launch a cutscene immediately when career
mode is selected from main menu. This avoids an edge case where the
player is in an area, and then selects a level for that area, and then
the player is shown a prologue for that area.