Refactor: Extracted StoryData from Global.
The player_spawn_id and sensei_spawn_id are specific to story mode and
don't need to be in Global. I've moved them into a new StoryData script under
PlayerData. They're currently not persisted, but they could be some day
if we want the player to continue from their previous location.