'bossio' cheat improvements.
'bossio' cheat now searches more diligently for a region with a boss
level. If none can be found, a 'cheat deactivated' sound plays.

'bossio' cheat now marks cutscenes as unviewed, and the boss level as
unplayed. This is slightly more future proof, just in case our logic
eventually changes to key off of whether or not the boss level has been
beaten. (It's currently based on the maximum distance travelled)