Cutscene renames: LevelHistory.add(), CutsceneManager, others
Renamed CutsceneManager to CutsceneQueue. Its behavior is really just
'here's a bunch of cutscenes to play before and after a level', so this
name is more specific.

Improved symmetry between LevelHistory's methods and ChatHistory's
methods. Simple methods names like add() and delete() are slightly
ambiguous, since this could involve deleting a single result, or
deleting an entire set of results. In fact, the methods were
inconsistent and delete DID delete an entire set of results. The names
now communicate that.