Extracted InvisibleObstacleTiler from ObstacleMap.
ObstacleMap's 'invisible obstacle' logic was previously implemented with an 'is
a' implementation, with ObstacleMapIndoors inheriting from ObstacleMap. By
redesigning this behavior with a 'has a' implementation, different ObstacleMaps
can pick and choose which autotiling behaviors they desire without running into
conflicts.

The 'invisible obstacle' logic was only preserved in
OverworldEnvironment and OverworldIndoorsEnvironment, because these are
the only environments used in free roam mode. The other environments
don't need hit detection because the player can't move.